Data.Machine.Plan

Description

Plans

type Plan k i o a = forall m. PlanT k i o m a

A Plan k i o a is a specification for a pure Machine, that reads inputs selected by k with types based on i, writes values of type o, and has intermediate results of type a.

A PlanT k i o a can be used as a PlanT k i o m a for any Monad m.

It is perhaps easier to think of Plan in its un-cps'ed form, which would look like:

 data Plan k i o a
   = Done a
   | Yield o (Plan k i o a)
   | forall z. Await (z -> Plan k i o a) (k i z) (Plan k i o a)
   | Fail

runPlan :: PlanT k i o Id a -> (a -> r) -> (o -> r -> r) -> (forall z. (z -> r) -> k i z -> r -> r) -> r -> r

Deconstruct a Plan without reference to a Monad.

newtype PlanT k i o m a

You can construct a Plan (or PlanT), turning it into a Machine (or MachineT).

Constructors

PlanT

Fields

runPlanT :: forall r. (a -> m r) -> (o -> m r -> m r) -> (forall z. (z -> m r) -> k i z -> m r -> m r) -> m r -> m r

yield :: o -> Plan k i o ()

Output a result.

await :: Plan (->) i o i

Wait for input.

await = awaits id

stop :: Plan k i o a

stop = empty

Handling multiple inputs

type Handle k i o = forall r. (o -> r) -> k i r

Many combinators are parameterized on the choice of Handle, this acts like an input stream selector.

 L :: Handle Merge (a,b) a
 R :: Handle Merge (a,b) b

type Fitting k k' o i = forall r. k o r -> k' i r

type Handle = Fitting (->)

awaits :: Handle k i j -> Plan k i o j

Wait for a particular input.

 awaits L  :: Plan Merge (a,b) o a
 awaits R  :: Plan Merge (a,b) o b
 awaits id :: Plan (->) i o i