Data.Machine.Source

Description

Sources

type Source b = forall m. Machine m b

A Source never reads from its inputs.

source :: Foldable f => f b -> Source b

Generate a Source from any Foldable container.

repeated :: o -> Source o

Repeat the same value, over and over.

cycled :: Foldable f => f b -> Source b

Loop through a Foldable container over and over.

cap :: Process a b -> Source a -> Source b

You can transform a Source with a Process.

Alternately you can view this as capping the Source end of a Process, yielding a new Source.

cap l r = l <~ r