Data.Machine.Stack

Description

Documentation

data Stack a r where

This is a simple process type that knows how to push back input.

Constructors

Push :: a -> Stack a ()

Pop :: Stack a a

stack :: Monad m => MachineT m k a -> MachineT m (Stack a) o -> MachineT m k o

Stream outputs from one Machine into another with the possibility of pushing inputs back.

peek :: Plan (Stack a) b a

Peek at the next value in the input stream without consuming it

pop :: Plan (Stack a) b a

Pop the next value in the input stream

push :: a -> Plan (Stack a) b ()

Push back into the input stream