Data.Machine.Tee

Description

Tees

type Tee a b = Machine ((->) a :+: (->) b)

A Machine that can read from two input stream in a deterministic manner.

data (m :+: n) a

The input descriptor for a Tee or TeeT.

Constructors

L (m a)

R (n a)

tee :: Process a a' -> Process b b' -> Tee a' b' c -> Tee a b c

Compose a pair of pipes onto the front of a Tee.

addL :: Process a b -> Tee b c d -> Tee a c d

Precompose a pipe onto the left input of a tee.

addR :: Process b c -> Tee a c d -> Tee a b d

Precompose a pipe onto the right input of a tee.

capL :: Source a -> Tee a b c -> Process b c

Tie off one input of a tee by connecting it to a known source.

capR :: Source b -> Tee a b c -> Process a c

Tie off one input of a tee by connecting it to a known source.