Data.Machine.Type

Description

Machines

newtype MachineT m k i o

A MachineT reads from a number of inputs and may yield results before stopping with monadic side-effects.

Constructors

MachineT

Fields

runMachineT :: m (Step k i o (MachineT m k i o))

data Step k i o r

This is the base functor for a Machine or MachineT.

Note: Machines are usually constructed from Plan, so this does not need to be CPS'd.

Constructors

Stop

Yield o r

forall t . Await (t -> r) (k i t) r

type Machine k i o = forall m. Monad m => MachineT m k i o

A Machine reads from a number of inputs and may yield results before stopping.

A Machine can be used as a MachineT m for any Monad m.

run :: MachineT Id k a b -> [b]

Run a pure machine and extract an answer.

runMachine :: MachineT Id k i o -> Step k i o (MachineT Id k i o)

runMachine = runId . runMachineT

encased :: Monad m => Step k i o (MachineT m k i o) -> MachineT m k i o

Building machines from plans

construct :: Monad m => PlanT k i o m a -> MachineT m k i o

Compile a machine to a model.

repeatedly :: Monad m => PlanT k i o m a -> MachineT m k i o

Generates a model that runs a machine until it stops, then start it up again.

repeatedly m = construct (forever m)

before :: Monad m => MachineT m k i o -> PlanT k i o m a -> MachineT m k i o

Evaluate a machine until it stops, and then yield answers according to the supplied model.

Reshaping machines

fitting :: Monad m => (forall a. k i a -> k' i' a) -> MachineT m k i o -> MachineT m k' i' o

Connect different kinds of machines.

fitting id = id

 fitting L :: Process a c -> Tee a b c
 fitting R :: Process b c -> Tee a b c
 fitting id :: Process a b -> Process a b

pass :: Functor (k i) => Handle k i o -> Machine k i o

Given a handle, ignore all other inputs and just stream input from that handle.

 pass id :: Process a a
 pass L  :: Tee a b a
 pass R  :: Tee a b b

stopped :: Machine k a b

This is a stopped Machine