Data.Machine.Unread

Description

Documentation

data Unread o a where

This is a simple process type that knows how to push back input.

Constructors

Unread :: a -> o -> Unread o a

Read :: (o -> a) -> Unread o a

peek :: Plan b (Unread a) a

Peek at the next value in the input stream without consuming it

unread :: a -> Plan b (Unread a) ()

Push back into the input stream