Data.Machine.Wye

Description

Wyes

type Wye a b = Machine (Y ((->) a) ((->) b))

A Machine that can read from two input stream in a non-deterministic manner.

data Y f g a

The input descriptor for a Wye or WyeT

Constructors

This (f a)

That (g a)

These (f a) (g a)

wye :: Process a a' -> Process b b' -> Wye a' b' c -> Wye a b c

Compose a pair of pipes onto the front of a Wye.

Precompose a Process onto each input of a Wye (or WyeT).

This is left biased in that it tries to draw values from the X input whenever they are available, and only draws from the Y input when X would block.

addX :: Process a b -> Wye b c d -> Wye a c d

Precompose a pipe onto the left input of a wye.

addY :: Process b c -> Wye a c d -> Wye a b d

Precompose a pipe onto the right input of a tee.

capX :: Source a -> Wye a b c -> Process b c

Tie off one input of a tee by connecting it to a known source.

capY :: Source b -> Wye a b c -> Process a c

Tie off one input of a tee by connecting it to a known source.