Data.Machine.Type

Description

Machines

newtype MachineT m k o

A MachineT reads from a number of inputs and may yield results before stopping with monadic side-effects.

Constructors

MachineT

Fields

runMachineT :: m (Step k o (MachineT m k o))

data Step k o r

This is the base functor for a Machine or MachineT.

Note: A Machine is usually constructed from Plan, so it does not need to be CPS'd.

Constructors

Stop

Yield o r

forall t. Await (t -> r) (k t) r

type Machine k o = forall m. Monad m => MachineT m k o

A Machine reads from a number of inputs and may yield results before stopping.

A Machine can be used as a MachineT m for any Monad m.

runT_ :: Monad m => MachineT m k b -> m ()

Stop feeding input into model, taking only the effects.

runT :: Monad m => MachineT m k b -> m [b]

Stop feeding input into model and extract an answer

run :: MachineT Identity k b -> [b]

Run a pure machine and extract an answer.

runMachine :: MachineT Identity k o -> Step k o (MachineT Identity k o)

runMachine = runIdentity . runMachineT

encased :: Monad m => Step k o (MachineT m k o) -> MachineT m k o

Pack a Step of a Machine into a Machine.

Building machines from plans

construct :: Monad m => PlanT k o m a -> MachineT m k o

Compile a machine to a model.

repeatedly :: Monad m => PlanT k o m a -> MachineT m k o

Generates a model that runs a machine until it stops, then start it up again.

repeatedly m = construct (forever m)

unfoldPlan :: Monad m => s -> (s -> PlanT k o m s) -> MachineT m k o

Unfold a stateful PlanT into a MachineT.

before :: Monad m => MachineT m k o -> PlanT k o m a -> MachineT m k o

Evaluate a machine until it stops, and then yield answers according to the supplied model.

preplan :: Monad m => PlanT k o m (MachineT m k o) -> MachineT m k o

Incorporate a Plan into the resulting machine.

Deconstructing machines back into plans

deconstruct :: Monad m => MachineT m k (Either a o) -> PlanT k o m a

tagDone :: Monad m => (o -> Bool) -> MachineT m k o -> MachineT m k (Either o o)

Use a predicate to mark a yielded value as the terminal value of this Machine. This is useful in combination with deconstruct to combine Plans.

finishWith :: Monad m => (o -> Maybe r) -> MachineT m k o -> MachineT m k (Either r o)

Use a function to produce and mark a yielded value as the terminal value of a Machine. All yielded values for which the given function returns Nothing are yielded down the pipeline, but the first value for which the function returns a Just value will be returned by a Plan created via deconstruct.

Reshaping machines

fit :: Monad m => (forall a. k a -> k' a) -> MachineT m k o -> MachineT m k' o

Connect different kinds of machines.

fit id = id

fitM :: (Monad m, Monad m') => (forall a. m a -> m' a) -> MachineT m k o -> MachineT m' k o

pass :: k o -> Machine k o

Given a handle, ignore all other inputs and just stream input from that handle.

pass id :: Process a a
pass L  :: Tee a b a
pass R  :: Tee a b b
pass X  :: Wye a b a
pass Y  :: Wye a b b
pass Z  :: Wye a b (Either a b)

starve :: Monad m => MachineT m k0 b -> MachineT m k b -> MachineT m k b

Run a machine with no input until it stops, then behave as another machine.

stopped :: Machine k b

This is a stopped Machine

stepMachine :: Monad m => MachineT m k o -> (Step k o (MachineT m k o) -> MachineT m k' o') -> MachineT m k' o'

Transform a Machine by looking at a single step of that machine.

Applicative Machines

class Appliance k where

An input type that supports merging requests from multiple machines.

Methods

applied :: Monad m => MachineT m k (a -> b) -> MachineT m k a -> MachineT m k b