Data.Machine.Type

Description

Machines

data Machine m a

A Machine reads from a number of inputs and may yield results before stopping.

Note: A Machine is usually constructed from Plan, so it does not need to be CPS'd.

(Now that we have a Monad of our own, this is no longer completely true.)

Constructors

Yield a (Machine m a)

forall i . Await (i -> Machine m a) (m i) (Machine m a)

Stop

run_ :: MonadPlus m => Machine m b -> m ()

run :: MonadPlus m => Machine m b -> m [b]

Building machines from plans

construct :: Plan o m a -> Machine m o

Compile a machine to a model.

repeatedly :: Plan o m a -> Machine m o

Generates a model that runs a machine until it stops, then start it up again.

repeatedly m = construct (forever m)

before :: Machine m o -> Plan o m a -> Machine m o

Evaluate a machine until it stops, and then yield answers according to the supplied model.

Reshaping machines

fit :: (forall a. m a -> n a) -> Machine m o -> Machine n o

Connect different kinds of machines.

fit id = id

pass :: m o -> Machine m o

Given a handle, ignore all other inputs and just stream input from that handle.

 pass id :: Process a a
 pass L  :: Tee a b a
 pass R  :: Tee a b b
 pass X  :: Wye a b a
 pass Y  :: Wye a b b
 pass Z  :: Wye a b (Either a b)