Data.Machine.Wye

Description

Wyes

type Wye a b c = Machine Y (a, b) c

A Machine that can read from two input stream in a non-deterministic manner.

type WyeT m a b c = MachineT m Y (a, b) c

A Machine that can read from two input stream in a non-deterministic manner with monadic side-effects.

data Y i c where

The input descriptor for a Wye or WyeT

Constructors

X :: Y (a, b) a

Y :: Y (a, b) b

Z :: Y (a, b) (Either a b)

wye :: Monad m => ProcessT m a a' -> ProcessT m b b' -> WyeT m a' b' c -> WyeT m a b c

Compose a pair of pipes onto the front of a Wye.

Precompose a Process onto each input of a Wye (or WyeT).

This is left biased in that it tries to draw values from the X input whenever they are available, and only draws from the Y input when X would block.

addX :: Monad m => ProcessT m a b -> WyeT m b c d -> WyeT m a c d

Precompose a pipe onto the left input of a wye.

addY :: Monad m => ProcessT m b c -> WyeT m a c d -> WyeT m a b d

Precompose a pipe onto the right input of a tee.

capX :: Monad m => SourceT m a -> WyeT m a b c -> ProcessT m b c

Tie off one input of a tee by connecting it to a known source.

capY :: Monad m => SourceT m b -> WyeT m a b c -> ProcessT m a c

Tie off one input of a tee by connecting it to a known source.